+++ Resources/MemberStats.cs
@@ -1,7 +1,7 @@
 ﻿namespace MailChimp.Resources {
     public class MemberStats : MailChimpResource
     {
-        public int AvgOpenRate { get; set; }
-        public int AvgClickRate { get; set; }
+        public float AvgOpenRate { get; set; }
+        public float AvgClickRate { get; set; }
     }
 }
